Create City With the Smallest Number of Neighbors at a Threshold Distance
There are n cities labeled from 0 to n-1 with m edges connecting them. Given the array edges where edges[i] = [from i , to i ,weight i]  represents a bidirectional and weighted edge between cities from i and to i, and given the integer distance Threshold. You need to find out a city with the smallest number of cities that are reachable through some path and whose distance is at most Threshold Distance. If there are multiple such cities, our answer will be the city with the greatest label.

Your Task:
You don't need to read input or print anything. Your task is to complete the function find City( ) which takes a number of nodes n, total number of edges m and vector of edges and distance Threshold. and return the city with the smallest number of cities that are reachable through some path and whose distance is at most Threshold Distance. If there are multiple such cities, return the city with the greatest label.