Create Minimum element in BST
Given the root of a Binary Search Tree. The task is to find the minimum valued element in this given BST.
Your Task:
The task is to complete the function minValue() which takes root as the argument and returns the minimum element of BST. If the tree is empty, there is no minimum element, so return -1 in that case.