Create 03-april-2024
Given a BST with n (n>=2) nodes, find the kth common ancestor of nodes x and y in the given tree. Return -1 if kth ancestor does not exist.
Nodes x and y will always be present in the input of a BST, and x != y.

Your task :
You don't have to read input or print anything. Your task is to complete the function kth Common Ancestor() that takes the root of the tree, k, x and y as input and returns the kth common ancestor of x and y.