Create 04-april-2024
Sum of all substrings of a number

Given an integer s represented as a string, the task is to get the sum of all possible sub-strings of this string.
As the answer will be large, return answer modulo 109+7. 

Your Task:
You only need to complete the function sumSubstrings that takes s as an argument and returns the answer modulo 109+7.