Create 06-april-2024
Count ways to N'th Stair

There are n stairs, and a person standing at the bottom wants to reach the top. The person can climb either 1 stair or 2 stairs at a time. Count the number of ways, the person can reach the top (order does not matter).

Your Task:
Your task is to complete the function countWays() which takes a single argument n and returns the answer.